Update auth layout #89
