--- apps/dailyscrum/src/app/app/(auth-pages)/layout.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-type Props = {
-  children: React.ReactNode;
-};
-
-const Layout = ({ children }: Props) => {
-  // TODO: work on auth layout UI
-  return children;
-};
-
-export default Layout;
